refactor: remove unnecessary Backend class; pass flag to tell if emcc bindings should be considered
