+++ backend/CMakeLists.txt
@@ -39,11 +39,9 @@ IF(${EMCC})
     set(CMAKE_CXX_COMPILER emcc)
     set(CMAKE_EXECUTABLE_SUFFIX ".js")
 
+    set(ENABLE_EMCC_BINDINGS "-DEMCC_COMPILER=1")
     set(EMCC_FLAGS "-sEXPORTED_RUNTIME_METHODS=['ccall'] -sEXPORT_ES6=1 -sMODULARIZE=1 -o MazeGenerator.js -sEXPORT_NAME='MazeGenerator' --bind ")
-    set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} ${EMCC_FLAGS}")
-
-    # Backend class is about emcc bindings
-    set(SRC ${SRC} src/Backend.cpp include/Backend.hpp)
+    set(CMAKE_CXX_FLAGS "${ENABLE_EMCC_BINDINGS} ${CMAKE_CXX_FLAGS} ${EMCC_FLAGS}")
 
     add_executable(mazeGenerator ${SRC})
     set_target_properties(mazeGenerator PROPERTIES LINK_FLAGS "-o mazeGenerator.html")
